refs is now a ptrdiff_t...
Rev: src/svalue.h:1.64

+++ src/svalue.h
@@ -5,7 +5,7 @@
 \*/
 
 /*
- * $Id: svalue.h,v 1.63 2000/07/29 06:31:07 hubbe Exp $
+ * $Id: svalue.h,v 1.64 2000/08/03 17:39:06 grubba Exp $
  */
 #ifndef SVALUE_H
 #define SVALUE_H
@@ -62,7 +62,7 @@ union anything
   struct object *object;
   struct program *program;
   struct pike_string *string;
-  INT32 *refs;
+  ptrdiff_t *refs;
   INT_TYPE integer;
   FLOAT_TYPE float_number;
   struct svalue *lval;   /* only used on stack */
